backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
python-dotenv
pydantic-settings
sqlalchemy
psycopg2-binary
python-jose[cryptography]
httpx
openai
pillow
pillow-heif
